Fixed minor bugs
 - fixed typos & wrong assignments
 - changed directory of temperory log file
 - revised tmpfile migration

+++ Daily/libupdate/cleanup.sh
@@ -2,7 +2,7 @@
 
 
 # clear potential terminal buffer
-sript -q /dev/null clear > /dev/null 2>&1 | tee /dev/null
+sript -q /dev/null tput clear > /dev/null 2>&1
 
 
 ################################################################################
@@ -27,20 +27,20 @@ logdate=$5
 
 # log file prepare
 logfile="/Library/Logs/Scripts/update/$logdate.log"
-tmpfile="/tmp/update.log"
+tmpfile="/tmp/log/update.log"
 
 
-# remove /tmp/update.log
+# remove /tmp/log/update.log
 rm -f $tmpfile
 
 
-# create /tmp/update.log & /Library/Logs/Scripts/update/logdate.log
+# create /tmp/log/update.log & /Library/Logs/Scripts/update/logdate.log
 touch $logfile
 touch $tmpfile
 
 
 # log current status
-echo "" >> $tmpfile
+
 echo "- /bin/bash $0 $@" >> $tmpfile
 
 
@@ -82,10 +82,13 @@ $logprefix echo | $logcattee | $logsuffix
 
 # archive caches if hard disk attached
 if [ -e /Volumes/Jarry\ Shaw/ ] ; then
-    # move caches
-    $logprefix echo "+ cp -rf cache archive $verbose $quiet" | $logcattee | $logsuffix
-    $logprefix cp -rf $cmd_v $(brew --cache) /Volumes/Jarry\ Shaw/Developers/ | $logcattee | $logsuffix
-    $logprefix echo | $logcattee | $logsuffix
+    # check if cache directory exists
+    if [ -e $(brew --cache) ] ; then
+        # move caches
+        $logprefix echo "+ cp -rf cache archive $verbose $quiet" | $logcattee | $logsuffix
+        $logprefix cp -rf $cmd_v $(brew --cache) /Volumes/Jarry\ Shaw/Developers/ | $logcattee | $logsuffix
+        $logprefix echo | $logcattee | $logsuffix
+    fi
 
     # if cask flag set
     if ( $arg_cask ) ; then
@@ -96,41 +99,66 @@ if [ -e /Volumes/Jarry\ Shaw/ ] ; then
 
     # if brew flag set
     if ( $arg_brew ) ; then
-        | $logcattee | $logsuffix echo "+ brew cleanup $verbose $quiet" | $logcattee | $logsuffix
+        $logprefix echo "+ brew cleanup $verbose $quiet" | $logcattee | $logsuffix
         $logprefix rm -rf $cmd_v $( brew --cache ) | $logcattee | $logsuffix
         $logprefix echo | $logcattee | $logsuffix
     fi
 fi
 
 
-# read /tmp/update.log line by line then migrate to log file
 while read -r line ; do
     # plus `+` proceeds in line
     if [[ $line =~ ^(\+\+*\ )(.*)$ ]] ; then
+        # add "+" in the beginning, then write to /Library/Logs/Scripts/update/logdate.log
         echo "+$line" >> $logfile
     # minus `-` proceeds in line
     elif [[ $line =~ ^(-\ )(.*)$ ]] ; then
+        # replace "-" with "+", then write to /Library/Logs/Scripts/update/logdate.log
         echo "$line" | sed "y/-/+/" >> $logfile
     # colon `:` in line
     elif [[ $line =~ ^([[:alnum:]][[:alnum:]]*)(:)(.*)$ ]] ; then
-        # log tag
-        prefix=`echo $line | sed "s/\[[0-9][0-9]*m//g" | sed "s/\(.*\)*:\ .*/\1/" | cut -c 1-3 | tr "[a-z]" "[A-Z]"`
-        # log content
-        suffix=`echo $line | sed "s/\[[0-9][0-9]*m//g" | sed "s/.*:\ \(.*\)*.*/\1/"`
+        # if this is a warning
+        if [[ $( tr "[:upper:]" "[:lower:]" <<< $line ) =~ ^([[:alnum:]][[:alnum:]]*:\ )(.*)(warning:\ )(.*) ]] ; then
+            # log tag
+            prefix="WAR"
+            # log content
+            suffix=`echo $line | sed "s/\[[0-9][0-9]*m//g" | sed "s/warning: //"`
+        # if this is an error
+        elif [[ $( tr "[:upper:]" "[:lower:]" <<< $line ) =~ ^([[:alnum:]][[:alnum:]]*:\ )(.*)(error:\ )(.*)$ ]] ; then
+            # log tag
+            prefix="ERR"
+            # log content
+            suffix=`echo $line | sed "s/\[[0-9][0-9]*m//g" | sed "s/error: //"`
+        # if this is asking for password
+        elif [[ $line =~ ^(Password:)(.*) ]] ; then
+            # log tag
+            prefix="PWD"
+            # log content
+            suffix="content hidden due to security reasons"
+        # otherwise, extract its own tag
+        else
+            # log tag
+            prefix=`echo $line | sed "s/\[[0-9][0-9]*m//g" | sed "s/\(.*\)*:\ .*/\1/" | cut -c 1-3 | tr "[:lower:]" "[:upper:]"`
+            # log content
+            suffix=`echo $line | sed "s/\[[0-9][0-9]*m//g" | sed "s/.*:\ \(.*\)*.*/\1/"`
+        fi
         # write to /Library/Logs/Scripts/update/logdate.log
         echo "$prefix: $suffix" >> $logfile
-    # error (red/[31m) line
-    elif [[ $line =~ ^(.*)(\[31m)(.*)$ ]] ; then
-        # add `ERR` tag and remove special characters then write to /Library/Logs/Scripts/update/logdate.log
-        echo "ERR: $line" | sed "s/\[[0-9][0-9]*m//g" >> $logfile
-    # warning (yellow/[33m)
-    elif [[ $line =~ ^(.*)(\[33m)(.*)$ ]] ; then
-        # add `WAR` tag and remove special characters then write to /Library/Logs/Scripts/update/logdate.log
-        echo "WAR: $line" | sed "s/\[[0-9][0-9]*m//g" >> $logfile
-    # other colourised `[??m` line
+    # colourised `[??m` line
     elif [[ $line =~ ^(.*)(\[[0-9][0-9]*m)(.*)$ ]] ; then
-        # add `INF` tag and remove special characters then write to /Library/Logs/Scripts/update/logdate.log
-        echo "INF: $line" | sed "s/\[[0-9][0-9]*m//g" >> $logfile
+        # error (red/[31m) line
+        if [[ $line =~ ^(.*)(\[31m)(.*)$ ]] ; then
+            # add `ERR` tag and remove special characters then write to /Library/Logs/Scripts/update/logdate.log
+            echo "ERR: $line" | sed "s/\[[0-9][0-9]*m//g" >> $logfile
+        # warning (yellow/[33m)
+        elif [[ $line =~ ^(.*)(\[33m)(.*)$ ]] ; then
+            # add `WAR` tag and remove special characters then write to /Library/Logs/Scripts/update/logdate.log
+            echo "WAR: $line" | sed "s/\[[0-9][0-9]*m//g" >> $logfile
+        # other colourised line
+        else
+            # add `INF` tag and remove special characters then write to /Library/Logs/Scripts/update/logdate.log
+            echo "INF: $line" | sed "s/\[[0-9][0-9]*m//g" >> $logfile
+        fi
     # empty / blank line
     elif [[ $line =~ ^([[:space:]]*)$ ]] ; then
         # directlywrite to /Library/Logs/Scripts/update/logdate.log
@@ -143,9 +171,9 @@ while read -r line ; do
 done < $tmpfile
 
 
-# remove /tmp/update.log
+# remove /tmp/log/update.log
 rm -f $tmpfile
 
 
 # clear potential terminal buffer
-sript -q /dev/null clear > /dev/null 2>&1 | tee /dev/null
+sript -q /dev/null tput clear > /dev/null 2>&1
